Modify installation process when vim_runtime exists

+++ install.sh
@@ -5,12 +5,12 @@ vimrtdir=~/.vim_runtime
 
 if [ ! -d ${vimrtdir} ]
 then
-    mv ~/.vimrc ~/.vimrc_bak
+    test -e ~/.vimrc && mv ~/.vimrc ~/.vimrc_bak
     git clone https://github.com/amix/vimrc.git ${vimrtdir} 
     git clone https://github.com/vim-scripts/taglist.vim ${vimrtdir}/sources_non_forked/vim-taglist
     git clone https://github.com/vim-scripts/wombat256.vim ${vimrtdir}/sources_non_forked/vim-wombat256
     mv ~/.vim_runtime/sources_non_forked/syntastic ${vimrtdir}/
-    sh ${vimrtdir}/install_awesome_vimrc.sh
 fi
 
+sh ${vimrtdir}/install_awesome_vimrc.sh
 cp -u my_configs.vim ${vimrtdir}/
